backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.3.0
gunicorn==20.1.0
numpy==1.24.2
pymongo==4.3.3
pytz==2022.7.1
redis==4.5.4
sqlparse==0.4.3
